requirements.txt
assemblyai
google-generativeai
hume-python
websockets>=11.0
pyaudio
